refactor: move RoomCleanupService to dedicated service package for better code hygiene
